+++ src/services/warpcast/index.ts
@@ -5,6 +5,7 @@ export { likeCast } from './like-cast'
 export enum HttpRequestMethod {
   GET = 'GET',
   POST = 'POST',
+  PATCH = 'PATCH',
   PUT = 'PUT',
   DELETE = 'DELETE',
 }